refactor: google/controller 패키지 구조를 controller/google 로 변경
